APID-40 - Upgraded the client and service to v16. Moved v15 service to src/Service/v15. Updated documentation and usage example.
