+++ example.php
@@ -10,7 +10,7 @@ use LBHounslow\Bartec\Service\BartecService;
 // BARTEC CLIENT USAGE
 $bartecClient = new BartecClient(
     new SoapClient(BartecClient::WSDL_AUTH),
-    new SoapClient(BartecClient::WSDL_COLLECTIVE_API_V15),
+    new SoapClient(BartecClient::WSDL_COLLECTIVE_API_V16),  // WSDL_COLLECTIVE_API_V15 also available
     'BARTEC_API_USERNAME',
     'BARTEC_API_PASSWORD'
 );
